// src/generated/client/model/getCatalogParams.ts
/**
 * Generated by orval v8.19.0 🍺
 * Do not edit manually.
 * Apimo Webservice API
 * Complete Apimo API documentation for real estate providers.
 * Access agencies, properties, contacts, leads and parameters.
 *
 * **Authentication** : HTTP Basic Auth with `provider_id:token`
 *
 * > ⚠️ **Note on error codes** : the documented HTTP codes (401, 403, 404, 422, 500)
 * > follow REST standard conventions. If in doubt, refer to the logs of your real requests.
 *
 * OpenAPI spec version: 1.0.0
 */

export type GetCatalogParams = {
/**
 * Label language (ISO 639-1)
 */
culture?: string;
};